# openapi.yaml
# Spin Trainer contract: source of truth (OpenAPI-first, ADR-0004). Changes from v0 and their reasons: ADR-0013.
openapi: 3.1.0
info:
  title: Spin Trainer API
  version: 0.2.0
  description: |
    API de Spin Trainer: catálogo de situaciones de Spin & Go (3-max y HU), rangos de referencia y personalizados,
    intentos del Quiz (eventos inmutables) y estadísticas agregadas.

    - Autenticación: JWT emitido por Supabase Auth (`Authorization: Bearer`), validado contra su JWKS.
    - Errores: Problem Details (RFC 9457, antes RFC 7807) con `type` URN `urn:spin-trainer:<tipo>`. Además de los
      estados declarados en cada operación, cualquiera puede responder 500 (`internal`) o 503 (`unavailable`: el
      emisor de los JWT no responde; reintentable) con el mismo formato. No se declaran operación a operación para que
      la validación contra la spec siga rechazando cualquier otro estado no previsto.
    - Trazabilidad: la cabecera opcional `X-Correlation-Id` se propaga a logs y se devuelve en la respuesta
      (se genera si no llega) y en `Problem.correlationId`.
    - Caché: `/situations` y `/ranges/default*` cambian solo con una migración de seed; responden con `ETag` y
      `Cache-Control: no-cache, private`, y con 304 si el `If-None-Match` coincide.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
tags:
  - name: situation
    description: Catálogo de las 16 situaciones (módulo situation).
  - name: range
    description: Rangos de referencia (PDF, seed Flyway) y personalizados del usuario (módulo range).
  - name: quiz
    description: Intentos del Quiz como eventos inmutables (módulo quiz).
  - name: stats
    description: Consultas agregadas sobre los intentos (módulo stats).

paths:
  /situations:
    get:
      tags: [situation]
      operationId: listSituations
      summary: Catálogo de situaciones, en orden de presentación
      responses:
        '200':
          description: OK
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Situation' } }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /ranges/default:
    get:
      tags: [range]
      operationId: listDefaultRanges
      summary: Todos los rangos de referencia cargados (una entrada por situación y stack con seed)
      description: Permite al cliente conocer qué combinaciones tienen rango y entrenar en modo aleatorio sin N peticiones.
      responses:
        '200':
          description: OK
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Range' } }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /ranges/default/{situation}/{stack}:
    parameters:
      - $ref: '#/components/parameters/situation'
      - $ref: '#/components/parameters/stack'
    get:
      tags: [range]
      operationId: getDefaultRange
      summary: Rango de referencia de una situación y stack
      responses:
        '200':
          description: OK
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Range' }
        '304': { $ref: '#/components/responses/NotModified' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  /ranges/user:
    get:
      tags: [range]
      operationId: listUserRanges
      summary: Todos los rangos personalizados del usuario
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Range' } }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /ranges/user/{situation}/{stack}:
    parameters:
      - $ref: '#/components/parameters/situation'
      - $ref: '#/components/parameters/stack'
    get:
      tags: [range]
      operationId: getUserRange
      summary: Rango personalizado del usuario
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Range' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Situación/stack desconocido o sin rango personalizado.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    put:
      tags: [range]
      operationId: putUserRange
      summary: Crea o reemplaza el rango personalizado (concurrencia optimista por versión)
      description: |
        `version` es obligatoria: `0` crea (409 si ya existe); `N` reemplaza la versión `N` (409 si cambió o se borró).
        Así una escritura nunca pisa otra sin saberlo. Las manos con la acción implícita se descartan al guardar.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RangeWrite' }
            example: { hands: { AA: MR_4B_C, KK: MR_4B_C, '72o': FOLD }, version: 0 }
      responses:
        '200':
          description: Reemplazado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Range' }
        '201':
          description: Creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Range' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [range]
      operationId: deleteUserRange
      summary: Borra el rango personalizado (vuelve a aplicar el de referencia). Idempotente.
      responses:
        '204': { description: Borrado o inexistente }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  /quiz/attempts:
    post:
      tags: [quiz]
      operationId: recordAttempt
      summary: Registra la respuesta a una mano del Quiz
      description: |
        El cliente envía solo lo que el usuario hizo. El servidor calcula `expected` contra el rango efectivo del usuario
        en ese momento (personalizado si existe, si no el de referencia; ADR-0012) y `correct`, y guarda de qué rango
        y versión salió (ADR-0006/0007). 422 si la situación/stack no tiene rango.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttemptWrite' }
            example: { situation: btn_open, stack: 25, hand: A5s, given: MR_F_F }
      responses:
        '201':
          description: Registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Attempt' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/NoRange' }
    get:
      tags: [quiz]
      operationId: listAttempts
      summary: Intentos del usuario, del más reciente al más antiguo (paginación por cursor)
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: cursor, in: query, description: Valor opaco `nextCursor` de la página anterior., schema: { type: string } }
        - { name: situation, in: query, schema: { $ref: '#/components/schemas/SituationKey' } }
        - { name: stack, in: query, schema: { $ref: '#/components/schemas/Stack' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttemptPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /stats/hands:
    get:
      tags: [stats]
      operationId: getHandStats
      summary: Intentos y aciertos agregados por situación, stack y mano
      description: |
        Agregación (GROUP BY) sobre los intentos. La política de estudio (manos difíciles, pesos, rankings)
        se aplica en el cliente sobre estas filas (ADR-0013).
      parameters:
        - { name: situation, in: query, schema: { $ref: '#/components/schemas/SituationKey' } }
        - { name: stack, in: query, schema: { $ref: '#/components/schemas/Stack' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/HandStat' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /stats/progress:
    get:
      tags: [stats]
      operationId: getProgress
      summary: Intentos y aciertos por día (solo días con actividad)
      parameters:
        - { name: days, in: query, schema: { type: integer, minimum: 1, maximum: 365, default: 30 } }
        - { name: tz, in: query, description: Zona horaria IANA para cortar los días., schema: { type: string, default: UTC, example: Europe/Madrid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ProgressDay' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }

  headers:
    ETag:
      description: Versión de la representación. Reenviarla en `If-None-Match` para recibir 304 si no cambió.
      schema: { type: string }

  parameters:
    situation:
      name: situation
      in: path
      required: true
      schema: { $ref: '#/components/schemas/SituationKey' }
    stack:
      name: stack
      in: path
      required: true
      description: Stack efectivo en BB, forma canónica sin ceros finales (`25`, `12.5`).
      schema: { $ref: '#/components/schemas/Stack' }

  responses:
    NotModified:
      description: Sin cambios respecto al `ETag` enviado en `If-None-Match`; el cliente reutiliza su copia.
    BadRequest:
      description: Petición inválida (urn:spin-trainer:validation), con detalle por campo en `errors`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Falta el JWT o no es válido (urn:spin-trainer:unauthorized).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Situación/stack desconocido o recurso inexistente (urn:spin-trainer:not-found).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: La versión enviada no es la actual (urn:spin-trainer:conflict).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NoRange:
      description: La situación/stack no tiene rango con el que corregir (urn:spin-trainer:no-range).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }

  schemas:
    SituationKey:
      type: string
      pattern: '^[a-z0-9_]+$'
      maxLength: 64
      example: btn_open
    Stack:
      type: number
      minimum: 1
      maximum: 100
      multipleOf: 0.5
      example: 12.5
    Hand:
      description: |
        Mano canónica del grid 13×13: pareja (`AA`) o rango alto + rango bajo + `s`/`o` (`AKs`, `T9o`).
        El patrón rechaza `AAs` o `AK`; el orden alto→bajo (`KAs` no es válido) lo valida el servidor (400).
      type: string
      pattern: '^(?:([AKQJT2-9])\1|([AKQJT2-9])(?!\2)[AKQJT2-9][so])$'
      example: AKs
    Action:
      type: string
      description: Acción compuesta. Una por mano y (situación, stack).
      enum: [MR_4B_C, MR_C_C, MR_C_F, MR_F_F, L_C_C, L_C_F, L_PUSH, L_F, ALLIN, 3BET, 3BET_C, CALL, CALL_VS_X2, ISO_C, ISO_F, LIMP, CHECK, FOLD]
      x-enum-varnames: [MR_4B_C, MR_C_C, MR_C_F, MR_F_F, L_C_C, L_C_F, L_PUSH, L_F, ALLIN, THREE_BET, THREE_BET_C, CALL, CALL_VS_X2, ISO_C, ISO_F, LIMP, CHECK, FOLD]
    Position:
      type: string
      enum: [BTN, SB, BB]
      description: En HU el botón es la SB.
    PriorAction:
      type: object
      required: [position, action]
      additionalProperties: false
      properties:
        position: { $ref: '#/components/schemas/Position' }
        action: { type: string, enum: [FOLD, LIMP, MIN_RAISE, RAISE, THREE_BET, CALL, SHOVE] }
    Situation:
      type: object
      required: [key, label, format, hero, priorActions, stacks, actions]
      additionalProperties: false
      properties:
        key: { $ref: '#/components/schemas/SituationKey' }
        label: { type: string }
        format: { type: string, enum: [3max, hu], description: Solo Spin & Go (ADR-0011). }
        hero: { $ref: '#/components/schemas/Position' }
        priorActions:
          type: array
          description: Acciones anteriores a la decisión del héroe, en orden. Vacía si abre el héroe.
          items: { $ref: '#/components/schemas/PriorAction' }
        stacks: { type: array, minItems: 1, uniqueItems: true, items: { $ref: '#/components/schemas/Stack' } }
        actions:
          type: array
          minItems: 2
          uniqueItems: true
          description: Acciones posibles. La implícita (manos no listadas) es FOLD, o CHECK si FOLD no está.
          items: { $ref: '#/components/schemas/Action' }
        notes: { type: string }
    Hands:
      type: object
      description: Mano → acción. Las manos ausentes tienen la acción implícita.
      propertyNames: { $ref: '#/components/schemas/Hand' }
      additionalProperties: { $ref: '#/components/schemas/Action' }
      maxProperties: 169
    Range:
      type: object
      required: [situation, stack, hands, source, version]
      additionalProperties: false
      properties:
        situation: { $ref: '#/components/schemas/SituationKey' }
        stack: { $ref: '#/components/schemas/Stack' }
        hands: { $ref: '#/components/schemas/Hands' }
        source: { type: string, enum: [default, user] }
        version:
          type: integer
          minimum: 1
          description: En `default`, versión del seed; en `user`, versión para concurrencia optimista.
        updatedAt: { type: string, format: date-time }
    RangeWrite:
      type: object
      required: [hands, version]
      additionalProperties: false
      properties:
        hands: { $ref: '#/components/schemas/Hands' }
        version: { type: integer, minimum: 0, description: 0 = crear; N = versión leída que se reemplaza. }
    AttemptWrite:
      type: object
      required: [situation, stack, hand, given]
      additionalProperties: false
      properties:
        situation: { $ref: '#/components/schemas/SituationKey' }
        stack: { $ref: '#/components/schemas/Stack' }
        hand: { $ref: '#/components/schemas/Hand' }
        given: { $ref: '#/components/schemas/Action' }
    Attempt:
      type: object
      required: [id, situation, stack, hand, given, expected, correct, rangeSource, rangeVersion, answeredAt]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        situation: { $ref: '#/components/schemas/SituationKey' }
        stack: { $ref: '#/components/schemas/Stack' }
        hand: { $ref: '#/components/schemas/Hand' }
        given: { $ref: '#/components/schemas/Action' }
        expected: { $ref: '#/components/schemas/Action' }
        correct: { type: boolean }
        rangeSource: { type: string, enum: [default, user] }
        rangeVersion: { type: integer, minimum: 1 }
        answeredAt: { type: string, format: date-time }
    AttemptPage:
      type: object
      required: [items, nextCursor]
      additionalProperties: false
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Attempt' } }
        nextCursor:
          type: [string, 'null']
          description: null en la última página.
          # Required even if null: the generated DTO always serializes it (the API omits the other nulls).
          x-field-extra-annotation: '@com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)'
    HandStat:
      type: object
      required: [situation, stack, hand, attempts, correct, lastAnsweredAt]
      additionalProperties: false
      properties:
        situation: { $ref: '#/components/schemas/SituationKey' }
        stack: { $ref: '#/components/schemas/Stack' }
        hand: { $ref: '#/components/schemas/Hand' }
        attempts: { type: integer, minimum: 1 }
        correct: { type: integer, minimum: 0 }
        lastAnsweredAt: { type: string, format: date-time }
    ProgressDay:
      type: object
      required: [date, attempts, correct]
      additionalProperties: false
      properties:
        date: { type: string, format: date }
        attempts: { type: integer, minimum: 1 }
        correct: { type: integer, minimum: 0 }
    Problem:
      type: object
      description: Problem Details (RFC 9457).
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference, example: 'urn:spin-trainer:validation' }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        correlationId: { type: string }
        errors:
          type: array
          description: Solo en 400. Un elemento por campo inválido.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string, example: 'hands.AAs' }
              message: { type: string }